Moved all models to one single folder
