+++ models/reviews.js
@@ -1,4 +1,4 @@
-const db = require("../../database/dbConfig.js");
+const db = require("../database/dbConfig.js");
 
 module.exports = {
   addReview,
