+++ models/reviews.test.js
@@ -1,12 +1,12 @@
-const db = require("../../database/dbConfig.js");
+const db = require("../database/dbConfig.js");
 const { prepTestingDB } = require("../../helpers/prepTestDB.js");
 
 const { addReview, getReviews, getReviewById } = require("./reviews.js");
 
-beforeAll(async() => {
-  prepTestingDB("user_reviews")
-  prepTestingDB("users")
-  await db.seed.run({ specific: '001-users.js' })
+beforeAll(async () => {
+  prepTestingDB("user_reviews");
+  prepTestingDB("users");
+  await db.seed.run({ specific: "001-users.js" });
 });
 
 const review1 = {
@@ -19,7 +19,7 @@ const review1 = {
   review:
     "No film before or since has embraced nihilism so willingly and made you enjoy it so much (hopefully, to some shame.) I'm not convinced there is any point to this movie but to be a movie, and that job it does quite well and provides a little context to Tarantino's later films.",
   tags: "",
-  watched_date: ""
+  watched_date: "",
 };
 
 const review2 = {
@@ -32,7 +32,7 @@ const review2 = {
     "The best of comedy, joy and wonder, distilled to perfection. It took me half an hour to wipe the smile off my face; this is a must-see classic.",
   tags: "",
   watched_date: "",
-  user_id: Number(2)
+  user_id: Number(2),
 };
 
 describe("letterboxd reviews model", () => {
@@ -61,9 +61,9 @@ describe("letterboxd reviews model", () => {
       expect.arrayContaining([
         expect.objectContaining({ name: "Reservoir Dogs" }),
         expect.objectContaining({ name: "Singin' in the Rain" }),
-        expect.objectContaining({ user_id: 2 })
+        expect.objectContaining({ user_id: 2 }),
       ])
-    )
+    );
   });
   // will continue to add tests and model functions if more functionality is needed.
 });
